외부 라이브러리 탑재 및 구조 변경

+++ Utility/GameTimer.cpp
@@ -97,7 +97,7 @@ float GameTimer::CheckTime()
 	__int64 currentTime;
 	QueryPerformanceCounter((LARGE_INTEGER*)&currentTime);
 
-	float totalTime = (currentTime - beforeTimePoint) * secondsPerFrame;
+	float totalTime = (float)((currentTime - beforeTimePoint) * secondsPerFrame);
 	beforeTimePoint = currentTime;
 
 	return totalTime;
